fix(wasm): enable rex protocol
