+++ internal/pkg/conn/p2p/network_native.go
@@ -33,7 +33,6 @@ import (
 	"github.com/rumsystem/quorum/internal/pkg/cli"
 	"github.com/rumsystem/quorum/internal/pkg/conn/pubsubconn"
 	"github.com/rumsystem/quorum/internal/pkg/options"
-	csdef "github.com/rumsystem/quorum/internal/pkg/storage/def"
 )
 
 var peerChan = make(chan peer.AddrInfo)
@@ -202,27 +201,6 @@ func NewNode(ctx context.Context, nodename string, nodeopt *options.NodeOptions,
 	}
 	go newnode.eventhandler(ctx)
 	return newnode, nil
-}
-
-func (node *Node) SetRumExchange(ctx context.Context, cs csdef.ChainStorageIface) {
-	peerStatus := NewPeerStatus()
-	var rexnotification chan RexNotification
-	rexnotification = make(chan RexNotification, 1)
-	var rexservice *RexService
-	rexservice = NewRexService(node.Host, peerStatus, node.NetworkName, ProtocolPrefix, rexnotification)
-	rexservice.SetDelegate()
-	rexchaindata := NewRexChainData(rexservice)
-	rexrelay := NewRexRelay(rexservice, cs)
-	rexservice.SetHandlerMatchMsgType("rumchaindata", rexchaindata.Handler)
-	rexservice.SetHandlerMatchMsgType("rumrelay", rexrelay.Handler)
-	networklog.Infof("Enable protocol RumExchange")
-
-	node.peerStatus = peerStatus
-	node.RumExchange = rexservice
-
-	if rexnotification != nil {
-		go node.rexhandler(ctx, rexnotification)
-	}
 }
 
 func (node *Node) Bootstrap(ctx context.Context, config cli.Config) error {
